refactor(#zimic): split cli tests by command
